+++ test/app/middleware/micropub/action.test.js
@@ -111,42 +111,6 @@ test('Undeletes a post', async t => {
   scope.done();
 });
 
-test.skip('Throws error undeleting if GitHub responds with an error', async t => {
-  // Mock GitHub create file request
-  const scope = nock('https://api.github.com')
-    .put(uri => uri.includes('baz.md'))
-    .replyWithError('Not found');
-
-  // Setup
-  store.set('https://foo.bar/baz.md', {
-    post: {
-      type: 'note',
-      path: 'baz.md'
-    },
-    mf2: {
-      type: ['h-entry'],
-      properties: {
-        content: ['Baz']
-      },
-      slug: ['baz']
-    }
-  });
-  const {app} = t.context;
-  const response = await app.post('/micropub')
-    .set('Accept', 'application/json')
-    .set('Authorization', `Bearer ${t.context.token}`)
-    .query({
-      action: 'undelete',
-      url: 'https://foo.bar/baz.md'
-    });
-
-  // Test assertions
-  t.is(response.status, 500);
-  t.is(response.body.error, 'error');
-  t.regex(response.body.error_description, /\bNot found\b/);
-  scope.done();
-});
-
 test.after(async () => {
   await fs.emptyDir(outputDir);
 });
