+++ test/app/middleware/micropub/create-post-attachment.test.js
@@ -1,10 +1,11 @@
+const path = require('path');
 const fs = require('fs-extra');
 const test = require('ava');
 const nock = require('nock');
 const request = require('supertest');
 
 const config = require(process.env.PWD + '/app/config');
-const outputDir = process.env.PWD + '/.ava_output/micropub-post';
+const outputDir = process.env.PWD + '/.ava_output/micropub-post-attachment';
 
 test.beforeEach(t => {
   config.data.dir = outputDir;
@@ -12,19 +13,25 @@ test.beforeEach(t => {
   t.context.token = process.env.TEST_INDIEAUTH_TOKEN;
 });
 
-test('Creates a post file', async t => {
+test('Creates a post file with attachment', async t => {
   // Mock request
   const scope = nock('https://api.github.com')
+    .persist()
     .put(/\b[\d\w]{5}\b/g)
-    .reply(201);
+    .reply(201)
+    .persist()
+    .put(/\b[\d\w]{5}\b/g)
+    .reply(200);
 
   // Setup
   const {app} = t.context;
+  const image = path.resolve(__dirname, 'fixtures/image.gif');
   const response = await app.post('/micropub')
     .set('Accept', 'application/json')
     .set('Authorization', `Bearer ${t.context.token}`)
-    .set('Content-Type', 'application/x-www-form-urlencoded; charset=utf-8')
-    .send('h=entry&content=Creates+a+post+file');
+    .field('h', 'entry')
+    .field('content', 'Creates a post file with attachment')
+    .attach('photo', image);
 
   // Test assertions
   t.is(response.status, 202);
