+++ test/app/middleware/micropub/create-post-error.test.js
@@ -4,7 +4,7 @@ const nock = require('nock');
 const request = require('supertest');
 
 const config = require(process.env.PWD + '/app/config');
-const outputDir = process.env.PWD + '/.ava_output/micropub-post';
+const outputDir = process.env.PWD + '/.ava_output/micropub-post-error';
 
 test.beforeEach(t => {
   config.data.dir = outputDir;
@@ -12,11 +12,11 @@ test.beforeEach(t => {
   t.context.token = process.env.TEST_INDIEAUTH_TOKEN;
 });
 
-test('Creates a post file', async t => {
+test('Throws error creating post if GitHub responds with an error', async t => {
   // Mock request
   const scope = nock('https://api.github.com')
     .put(/\b[\d\w]{5}\b/g)
-    .reply(201);
+    .replyWithError('not found');
 
   // Setup
   const {app} = t.context;
@@ -24,12 +24,12 @@ test('Creates a post file', async t => {
     .set('Accept', 'application/json')
     .set('Authorization', `Bearer ${t.context.token}`)
     .set('Content-Type', 'application/x-www-form-urlencoded; charset=utf-8')
-    .send('h=entry&content=Creates+a+post+file');
+    .send('h=entry&content=Throws+error+if+GitHub+responds+with+an+error');
 
   // Test assertions
-  t.is(response.status, 202);
-  t.is(response.body.success, 'create_pending');
-  t.regex(response.header.location, /\b[\d\w]{5}\b/g);
+  t.is(response.status, 500);
+  t.is(response.body.error, 'error');
+  t.regex(response.body.error_description, /\bnot found\b/);
   scope.done();
   nock.cleanAll();
 });
